Chat room list component divided in chatroom/dm header and chatroom/dm list
