--- front/component/Chat/ChatListLayout.tsx
@@ -1,3 +1,0 @@
-export default function ChatListLayout() {
-  return <div></div>;
-}
