+++ front/component/Chat/RoomList/DmHeader.tsx
@@ -1,0 +1,16 @@
+export default function DmHeader() {
+  return (
+    <div>
+      <h1>DM</h1>
+      <hr />
+      <style jsx>{`
+        h1 {
+          font-family: "Fragment Mono", monospace;
+          font-size: 25px;
+          font-weight: bold;
+          margin-left: 10px;
+        }
+      `}</style>
+    </div>
+  );
+}
